Remove duplication of error message parsing
