+++ apps/sdp-web/src/app/dashboard/custody/wallet-activity.data.ts
@@ -4,7 +4,7 @@ import type {
   TokenTransactionListItem,
 } from "@sdp/types";
 import type { SdpApiClient } from "@/lib/sdp-api";
-import { readTransactionParam, toTitleCase } from "../activity-format-utils";
+import { parseErrorMessage, readTransactionParam, toTitleCase } from "../activity-format-utils";
 import { type FetchResult, fetchPaymentTransfers } from "../payments/payments-page.data";
 
 export type WalletActivitySourceKind = "payments" | "issuance";
@@ -36,18 +36,6 @@ interface DashboardWalletActivityEnvelope {
 }
 
 const WALLET_ACTIVITY_LIMIT = 20;
-
-function parseErrorMessage(body: string): string {
-  try {
-    const parsed = JSON.parse(body) as {
-      error?: { message?: string };
-      message?: string;
-    };
-    return parsed?.error?.message ?? parsed?.message ?? body;
-  } catch {
-    return body || "Unknown error";
-  }
-}
 
 function resolvePaymentOperation(transfer: PaymentTransferSummary): string {
   if (transfer.direction === "inbound") {
